refactor: DTO Valid 추가, 이름 변경
